requirements.txt
fastapi
uvicorn
groq
modal
python-dotenv
instructor